+++ routes/taxonomies/actions/index.ts
@@ -1,4 +1,2 @@
-export { default as activateAction } from './activate';
-export { default as deactivateAction } from './deactivate';
-export { default as deleteTaxonomyAction } from './delete';
-export { default as editTaxonomyAction } from './edit';
+export { useEditTaxonomyAction } from './edit';
+export { default as quickEditTaxonomyAction } from './quick-edit';